refactor: 백엔드 기능 단위 파일/폴더 분리 (#44)
