+++ backend/src/main/java/com/costwise/workflow/ApprovalWorkflowService.java
@@ -1,7 +1,9 @@
-package com.costwise.service;
+package com.costwise.workflow;
 
-import com.costwise.api.dto.ApprovalWorkflowResponse;
+import com.costwise.api.dto.workflow.ApprovalWorkflowResponse;
+import com.costwise.audit.AuditLogService;
 import com.costwise.api.dto.PortfolioSummaryResponse;
+import com.costwise.service.PortfolioSummaryService;
 import java.time.LocalDateTime;
 import java.util.LinkedHashMap;
 import java.util.List;
