+++ backend/src/test/java/com/costwise/api/audit/AuditLogControllerTest.java
@@ -1,4 +1,4 @@
-package com.costwise.api;
+package com.costwise.api.audit;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -18,6 +18,7 @@ import java.util.Base64;
 import java.util.List;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
+import com.costwise.api.support.JsonFieldReader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
