+++ backend/src/test/java/com/costwise/api/support/JsonFieldReader.java
@@ -1,15 +1,15 @@
-package com.costwise.api;
+package com.costwise.api.support;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-final class JsonFieldReader {
+public final class JsonFieldReader {
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
     private JsonFieldReader() {}
 
-    static String read(String json, String fieldName) throws Exception {
+    public static String read(String json, String fieldName) throws Exception {
         JsonNode root = OBJECT_MAPPER.readTree(json);
         JsonNode field = root.get(fieldName);
         if (field == null || field.isNull()) {
