+++ backend/src/test/java/com/costwise/workflow/ApprovalWorkflowServiceTest.java
@@ -1,9 +1,12 @@
-package com.costwise.service;
+package com.costwise.workflow;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.costwise.api.dto.ApprovalWorkflowResponse;
+import com.costwise.api.dto.workflow.ApprovalWorkflowResponse;
+import com.costwise.audit.AuditLogRepository;
+import com.costwise.audit.AuditLogService;
+import com.costwise.service.PortfolioSummaryService;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import java.util.ArrayList;
 import java.util.Comparator;
