+++ docs/dev-logs/2026-04-22-backend-feature-split.md
@@ -1,0 +1,39 @@
+# 2026-04-22 Backend Feature Split
+
+## 대상
+
+- 브랜치: `feat/44-slice`
+- 범위: 이슈 `#44` 백엔드 기능 단위 파일/폴더 분리
+
+## 기준
+
+- 컨트롤러/서비스/저장소/DTO 책임을 기능 단위로 분리
+- 동작 회귀 없이 패키지 경계만 명확화
+- 테스트 구조를 구현 구조와 정렬
+
+## A/B 비교
+
+### A. 기존 패키지 유지 + 파일명만 정리
+
+- 장점: 변경량이 작다.
+- 단점: 책임 경계가 여전히 모호하다.
+
+### B. 기능 패키지 중심 재배치
+
+- 장점: 유지보수성과 영향 범위 예측성이 좋아진다.
+- 단점: import/테스트 경로 수정량이 커진다.
+
+## 선택
+
+- `B`를 선택했다.
+- 이유: 이슈 목표(기능 단위 구조 확립)에 직접 부합한다.
+
+## 검증
+
+- `./gradlew.bat check` 통과
+- 기능별 핵심 테스트 스위트 통과(에이전트 실행 결과 + 메인 재검증)
+
+## 리스크
+
+- 대규모 이동 리팩터링이라 이후 병렬 브랜치와 충돌 가능성이 높다.
+- 후속 작업에서 패키지 import drift 점검이 필요하다.
